Foundations for a workthrough
- Preface added to the readme to highlight the purpose of any changes made.
- CMakeLists updated with min Cpp version (only required for contiguous brackets).
- Various variables renamed
- Some detailed noted on analysis, which could do with a few more passes.

The approach to analysis, and specifically resolution, is interest.
And, is quite different to the resolution buffer I use.

At a general level, it seems the appraoches make different tradeoffs.
With MiniSAT the clause grows, and so many require mallocs (though the custom allocator may mitigate this) while with otter_sat the persistent buffer is a somewhat large overhead, though unless I overlooked the implementation, should only require a single alloc for the extracted clause.
And, though this overhead is relatively large, I don't think it's so large as to be an issue, as even with a large number of atoms it should be easy to handle.

Further, MiniSAT requires upholding the watch invariant.
I'd like to avoid requiring this in general, as at the moment a relatively small change to BCP could allow clauses to be constant and open up some possibilities for parallel work.
Though, on the other hand given use of the invariant is made, a test of the MiniSAT implementation and then making the invariant an option if worthwhile could be the way to go…