Additional notes and adjustments
Notes on adding clauses have been made.
And, singular functions for adding short clauses have been removed, given these have no real use.
E.g. there is no internal use for adding ternary clauses as a triple of literal parameters, and otherwise the clause can be added as a vec (with very slight overhead).

An note on the issue with calculating nFreeVars has been made, which I think identifies the issue referenced in a past note.
No fix is given, as it's not very important, and would likely be a headache without an additional counter (which would in turn trivialise things).