Minor notes
Examination of the DIMACs code left some puzzlment, but after realising the note on a need to rewrite, perhaps that's to be expected.
The code itself is fine, but the purpose is unclear to me.

Elsewhere, some variables renamed for clarity and some further convenience methods for assuming a small collection of variables removed.